refactor(runtime): dispatch exec-input surface splits to its module (F-O4)
The 1500-LOC wall fired on dispatch.rs (1502 · the env threading crossed
it): the pure input-shaping unit — build_exec_input (argv per element ·
shell whole), render_exec_io (cwd/env/stdin), the shell leading-program
identity, and the capture enum bridge — moves to dispatch/exec_io.rs
with its four render tests (the dispatch/{permits,regate,sandbox}
precedent). The exec DISPATCH itself (gates · re-gate · spawn) stays in
the parent · 1299 lines. cargo test -p nika-runtime --lib 240 green ·
clippy -D warnings clean.
